refactor(session): extract context hint builders
